refactor: remove render script, separate mesh renderer
