+++ package.json
@@ -3,8 +3,8 @@
   "browser": "dist/browser.js",
   "module": "dist/module.js",
   "dependencies": {
-    "@oasis-engine/controls": "^0.2.0",
-    "@oasis-engine/stats": "^0.2.0"
+    "@oasis-engine/controls": "^0.2.2",
+    "@oasis-engine/stats": "^0.2.2"
   },
   "devDependencies": {
     "@ali/ci": "latest",
@@ -19,7 +19,7 @@
     "@rollup/plugin-babel": "^5.2.1",
     "@rollup/plugin-commonjs": "^16.0.0",
     "@rollup/plugin-node-resolve": "^10.0.0",
-    "oasis-engine": "^0.2.0",
+    "oasis-engine": "^0.2.2",
     "rollup": "^2.32.0",
     "rollup-plugin-terser": "^7.0.2",
     "vite": "1.0.0-rc.5",
